Move printing commands to output.rs
